requirements.txt
opencv-python>=4.8.0
ultralytics>=8.0.0
easyocr>=1.7.0
flask>=2.3.0
flask-cors>=4.0.0
numpy>=1.24.0
torch>=2.0.0
Pillow>=10.0.0
gunicorn>=21.2.0

